requirements.txt
django==4.0.6
psycopg2-binary==2.9.3
requests==2.28.1
pillow==9.2.0
djangorestframework==3.13.1

pytest>=7.1.2
pytest-django>=4.5.2
factory-boy>=3.2.1
pytest-factoryboy>=2.5.0
faker>=13.15.0
django-filter>=22.1
scrapy==2.6.2
pycryptodome==3.15.0
